Dealt with Warnings on Compiling MIDI Plugin
1. -Wsign-conversion (1 Warning)

Compiler: clang version 10.0.0-4ubuntu1
Target: x86_64-pc-linux-gnu
Note: 10.0.0-4ubuntu1 issues 1 warning in juce_LV2_Wrapper.cpp.

2. -Wsign-conversion (3 Warnings)

Compiler: Raspbian clang version 11.0.1-2+rpi1
Target: armv6k-unknown-linux-gnueabihf
Note: 11.0.1-2+rpi1 issues 3 warnings in juce_LV2_Wrapper_Exporter.cpp, and 1 warning in the same way as 10.0.0-4ubuntu1.

+++ modules/juce_audio_plugin_client/LV2/juce_LV2_Wrapper.cpp
@@ -1169,7 +1169,7 @@ public:
                         if (event->body.type == uridMidiEvent)
                         {
                             const uint8* data = (const uint8*) (event + 1);
-                            midiEvents.addEvent (data, event->body.size, static_cast<int> (event->time.frames));
+                            midiEvents.addEvent (data, static_cast<int> (event->body.size), static_cast<int> (event->time.frames));
                             continue;
                         }
  #endif
